# Conflicts:
#	app/src/main/res/drawable/earned_points.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/teamx/equiz/data/models/loginData/User.kt
@@ -21,7 +21,7 @@ data class User(
     val street: String,
     val uniqueID: String,
     val updatedAt: String,
-    val wallet: Int,
+    val wallet: Double,
     val wishlist: List<Wishlist>,
     val zip: String
 )
